requirements.txt
# Core Dependencies
python-dotenv>=0.19.0
pymongo>=4.0.0
pydantic>=1.9.0
motor>=3.3.1
requests>=2.31.0
beautifulsoup4>=4.12.2
playwright>=1.36.0
numpy>=1.26.2
pandas>=2.0.0
matplotlib>=3.8.2
seaborn>=0.13.0
scikit-learn>=1.3.2
nltk>=3.8.1
spacy>=3.7.2
wordcloud>=1.9.3
Pillow>=10.1.0
python-dateutil>=2.8.2
pytz>=2023.3
tqdm>=4.66.1
click>=8.1.7
rich>=13.7.0
tabulate>=0.9.0
validators>=0.22.0
urllib3>=2.1.0
certifi>=2023.11.17
charset-normalizer>=3.3.2
idna>=3.6
setuptools>=69.0.3
wheel>=0.42.0

# Development Dependencies
pytest>=7.0.0
pytest-asyncio>=0.20.0
black>=22.8.0
flake8>=5.0.0
mypy>=0.990
types-python-dateutil>=2.8.19
types-requests>=2.28.11



playwright-stealth==1.0.2

# Testing
pytest
pytest-asyncio

